Create a fixpoint combinator for graph fixers (#1425)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1425

A graph fixer is a function which mutates a graph and produces a `(bool, errors)` pair; the bool indicates whether the fixer did any mutation, and the errors indicate whether there was a fatal error detected.

I've added a fixpoint combinator that takes a graph fixer and produces the graph fixer that keeps running the input fixer until either it stops making changes or produces any error.

With this combinator we now run the initial arithmetic rewriting pass over and over again until it either errors or stops finding things to fix.  This then means that we can do the rewrites in *any order*. This then eliminates the "chicken and egg" issues in the graph fixing step, where we had fixers that must run both before and after other fixers.

To demonstrate that we've achieved this desirable property I've reordered the logsumexp node fixer and the multiary addition node fixer. Previously we required that the logsumexp fixer run after the multiary addition fixer, but now it does not matter which order they run in because we will keep re-running the pass until we stop making progress.

Reviewed By: yucenli

Differential Revision: D34700882

fbshipit-source-id: 788936b29920fdcb12ae2cc2f43f89f9ad6afa0e