Refactor unsupported node error pass
Summary:
I want to make it easier to write passes that *never* rewrite a node and only produce errors; I'm doing that by making a new higher-order function. Its input is a function from graph *edges* to *errors*; its output is a graph fixer.  (Recall that a graph fixer is a function with signature `()->(bool, errors)` where the bool indicates whether the pass mutated the graph.

How do we represent an edge? An edge in a DAG has two ends which we'll call parent and child; each child has a list of its parents. We pass in the child and the index into its parent list to represent the edge.

In this diff I add the new higher-order function and rewrite the unsupported node detector to use the new mechanism. I'll add other error passes in upcoming diffs.

Reviewed By: yucenli

Differential Revision: D34763089

fbshipit-source-id: 41bb607d8527d9d270996d769de9fba7ca302623