Fix random walk adaptation for scalar latent random variable (#1434)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1434

Fix https://github.com/facebookresearch/beanmachine/issues/1414

The main change here is to make sure that the scalar is not 0-dimensional before attempting to check the size of its first dimension, so that we won't run into index out of range error during adaptation. I also updated one of the test to use a model with scalar latent variable to make sure that the fix is working as expected.

One interesting I notice while fixing the adaptation is that the "adaptive random walk" test wasn't testing the adaptation at all. Instead of running 30 adaptive iterations, it was set to sample 30 chains (but still manage to pass the test). And then, after fixing the inference args, the inference began to fail 😂. It turned out that it was because we didn't cap the acceptance probability, which, as in the case of the failing test, can result in the step size being scaled up to infinity.

{F728916681}

Reviewed By: jpchen

Differential Revision: D36121999

fbshipit-source-id: 874939eb07846edebe6b10d0d36e3efd814cfe04