Compiler now detects invalid matrix multiplications (#1437)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1437

BMG requires that a matrix multiplication have the left operand's column count exactly equal the right operand's row count; here we enforce that rule in the compiler by adding a new error checking pass that runs after the unsupported node checking pass.

We do not yet enforce the BMG requirement that the operands be real, negative real, positive real or probability matrices; I'll add that code in an upcoming diff.

While I was adding a new error pass I took this opportunity to add a pass which should help me or other future developers detect mistakes early. By the time we get past the unsupported node checking pass, every node remaining in the graph (except constants, which are type checked during edge requirements checking) should have a type assigned to it by the BMG lattice typer.

If that invariant is not met then the upcoming passes can crash. Rather than allowing that crash and debugging the problem then, it will be easier for all involved if we detect the problem early and do not run passes which require an invariant that we know to be violated.

Reviewed By: yucenli

Differential Revision: D34775673

fbshipit-source-id: 0b26504a2c72309619fa1754323aa8d0cd2d24d9