Compiler now requires matmul operands to be real matrices
Summary:
This diff completes my efforts to finish the implementation of matrix multiplication in the compiler. We now ensure that the inputs to any matrix multiplication are double-valued matrices, as required by BMG.

BMG requires that the inputs to a matrix multiplication be a matrix of doubles; whether reals, negative reals, positive reals or probabilities, it does not care so long as they are not naturals or bools. To ensure this:

* Constant matrices that are inputs are automatically emitted as double-valued matrices even if their elements are naturals.

* stochastic inputs that are not real, negative real, positive real or probability matrices have a to-real-matrix operator inserted on the edge.

Reviewed By: yucenli

Differential Revision: D34865292

fbshipit-source-id: 0355e93a535251bb3a4a79acb933f4dd6cc1aa34