Start implementing tensor sum operator support in compiler
Summary:
We wish to compile Bean Machine models containing tensor sum operators into BMG graphs; this diff gets a start on that feature by detecting usages of the sum operator on stochastic quantities and giving an error message saying that the operation is not yet supported.

In upcoming diffs we will add a transformation pass which turns sum nodes into multiary additions.

Reviewed By: yucenli

Differential Revision: D34905869

fbshipit-source-id: 9b89daed7b3b820f4c824d0c93a9d386dd07d0f2