Improve error detection of models using unsupported stochastic operations
Summary:
The compiler should detect when a model uses an operation which we know for certain cannot be accumulated into the graph and immediately throw a helpful error message.   This diff resolves a number of scenarios where we did not do so, and instead crashed somewhere in the compiler.

The implementation choice here is to make *every* function in the torch module and any submodule a "special function"; that is, either we know how to represent that function as a graph node (`log`, `exp`, `add`...) or we do not. If we do not then we say so, rather than what we did before, which was either crash upon fetching the attribute, or crash upon calling the function with a graph node as an argument.

Reviewed By: wtaha

Differential Revision: D35097003

fbshipit-source-id: 8e80856ef880f1d73aec854f701f6c7ffa305b5d