Add compiler support for five more torch operators
Summary:
BMG has no nodes corresponding to the torch log2, log10, log1p, exp2, and sqrt operators, but it does not need them. We now detect usages of these five operators in models and translate them to:

    log2(x)  --> log(x) * (1/log(2))
    log10(x) --> log(x) * (1/log(10))
    log1p(x) --> log(1+p)
    exp2(x)  --> 2**x
    sqrt(x)  --> x**0.5

all of which we can represent in BMG.

Note that we might want to add a log1p operator to BMG because a dedicated implementation of log(1+p) can be made more accurate for certain values of p.  (Namely, if p is a probability very close to zero then log1p(-p) is the log probability of the complement of p, which is an operation we do in some models, and it's nice to get this as accurate as possible.)

Reviewed By: wtaha

Differential Revision: D35124235

fbshipit-source-id: f90850b81f95110e13d70bab6b6631b17feb4480