Fix bug in multiplication devectorizer
Summary:
The multiplication devectorizer is intended to find models which contain element-by-element tensor-style multiplication of 1-d or 2-d tensors, and decompose them into individual multiplications. (Because BMG does not support tensor multiplications.)  However, we have an operator in BMG for matrix-scalar multiplication. The intention of the code here was "do not devectorize multiplication if either operand is a scalar" but I accidentally wrote "do not devectorize if either operand is a 1-d vector or scalar", which is plainly wrong.

I've fixed the bug and created a test case which demonstrates that we do the devectorization correctly.

Reviewed By: wtaha

Differential Revision: D35264508

fbshipit-source-id: 43aaa2ae0f5b99c8483d9acd3b581a9dfcaf7e09