Fix two small bugs in devectorizer logic (#1452)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1452

While testing the log1p operator added in a previous diff I noticed that the devectorizer was doing a bad job on this operator. This led me to discover a number of bugs in the devectorizer; in this diff I fix two small ones and in the next diff I do a lot more work to make the devectorizer more robust.

One bug fixed here is a result of an unfortunate mismatch between the torch type system and the BMG type system that has led to several bugs. The underlying issue is that in torch, tensors are "row major" and in BMG, matrices are "column major", so if we have a `Size([2, 3])` tensor in the original model, it will have two rows and three columns in the original model, but three rows and two columns in the final BMG graph.

We have code paths in the graph transformation pass which need to know the original torch size AND the final BMG dimensions, and therefore it is easy to accidentally use one when you mean to use the other; the `_size_to_matrix` function is supposed to compute the *torch* equivalent size of a *BMG* `TO_MATRIX` node, but I forgot to swap the size back to use the torch convention.

The second bug is due to an oddity in how BMG deals with indexing into a matrix.  All matrices in BMG are two dimensional and there are two indexing operators.  The column indexer takes a matrix and returns a one-column matrix; the vector indexer takes a one-column matrix and returns an atomic value. The bug was that in some cases I would generate too many index nodes for a one-column matrix.

While debugging these issues I had some difficulty understanding my own code because when I refactored the problem fixing passes in a previous diff in this stack, I chose poor names for nested functions; those functions ended up on the call stack and it was difficult to then see what operations they corresponded to. As part of this bug fix I also renamed some of those nested functions so as to more easily understand a call stack while debugging.

Testing this operator in a vectorized model revealed a larger bug in the design and implementation of the devectorizer, which I'll fix in the next diff.

Reviewed By: wtaha

Differential Revision: D35325694

fbshipit-source-id: c3250bb9f4dd34f2b7f8b721f843b36e2f18de3c