Fix shift operator test case (#1454)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1454

Torch has changed the semantics of the `<<` and `>>` operators on tensors to ONLY work on integer-valued tensors, not float valued tensors. Because of this breaking change our test case which verified that the operators work on floats stopped working.

I've updated the test to use only integers and it now passes.

Reviewed By: gafter, ToddSmall

Differential Revision: D36527042

fbshipit-source-id: 669ba54a4298b11407edafb93b00f5136d6a06f1