Add stronger preconditions to devectorizer
Summary:
Many Bean Machine models are vectorized -- a model for instance might sample from a `Normal(tensor([0., 0]), tensor([1., 1]))` distribution to get a two-element vector of samples, and then do some operation on that vector; take its log or add it to another vector, or whatever.

Though BMG does have an IID_SAMPLE operator, we do not make use of it. And even if we did, (1) it does not work with all inference methods, (2) it doesn't help if the samples in the vector are not identically distributed, and (3) we still need to deal with compilation of models that perform other operations on 1- and 2-dimensional tensors.

The devectorizer is a graph-fixing pass that adds O(n) extra nodes to the graph to break up vectorized operations into n operations on atomic values. My initial implementation of the devectorizer was done for the CLARA model sprint and was rather hastily written; as we are now running more complex models through it, we're finding some rather fundamental bugs.

Suppose we have a graph of this form

     SOURCE  -->   OP   -->   SINK

where SOURCE produces a vector, OP takes an atomic value and produces an atomic value, and SINK takes a vector.  That's not going to work, but we can introduce O(n) additional nodes for a vector of size n:

          /-->  INDEX(0) --> OP -->
    SOURCE -->  INDEX(1) --> OP -->  TO_MATRIX --> SINK
          \-->  INDEX(2) --> OP -->
           ...

And now everyone is happy.  However my original implementation had some problems:

* Most importantly: the fixer must NOT introduce index nodes onto source nodes that are not indexable! The code assumed that any source node that had a tensor shape with more than one element could have an index applied to it, but that's not the case. BMG only allows certain nodes to be indexed.

This came up when I was experimenting with the vectorized `log1p` operator mentioned in the previous diff. We need to ensure that a vectorized `log1p(x)` is rewritten into `log(add(1, x))` operator FIRST, and then the log and add operators are devectorized normally.  In the buggy scenario, we introduced index nodes directly on the log1p node and things went downhill from there.

We now have an allow-list of the node types on which we allow the devectorizer to insert an outgoing edge to an index node.

* Second, the code is now much more explicit about the requirements on the inputs for a node to be devectorized. The inputs must ALL be either indexable or scalar, and AT LEAST ONE input must be indexable and of size more than one.

* Third, we have a special case for multiplication of a scalar by a vector; rather than devectorizing it we turn it into a scalar multiplication node in a different fixer, so we skip that here.

* Fourth, the code is still not correct for multiary additions and multiplications; I will fix that in another diff.  But it is slightly more correct now insofar as we now skip devectorizing multiary multiplications where at least one operand is a scalar and at least one is a vector.

* Fifth, I noticed that I forgot to compute the size of a matrix scale node.

* Sixth, it was poorly commented. I've added a number of explanatory comments to this tricky code.

Reviewed By: yucenli

Differential Revision: D35419360

fbshipit-source-id: 37baacf8872da3b7b0b5ec4fbadb92aaac17b3e9