bug fix for matrix operators
Summary:
Constant nodes in the graph do not have an initialized `Grad1` and `Grad2`, since these gradients are never necessary for taking the gradient of the log prob w.r.t to random variables. However, we were assuming that the Grads were always initialized fully. This caused BMG to crash in cases like
"constant_matrix MATRIX_ADD sample_matrix".

I have updated the gradient computations for `MATRIX_ADD` and `MATRIX_MULTIPLY` to account for this.

Reviewed By: shannyang

Differential Revision: D36889907

fbshipit-source-id: 132650f1bff7bf12ca889bc69e3028f51c497362