Convert inline HTML style attributes
### Motivation

We currently do not handle HTML in Markdown as any style attributes will
not be formatted correctly for React. This change uses `mdformat` in
combination with finding inline style attributes that are reformatted so
React can consume them.

The change allows a user to write HTML in a Jupyter session with inline
styles applied to any HTML tag they wish. Those styles will be converted
to a React appropriate object that Docusaurus can render correctly. Thus
allowing the user to see the same styles being applied to both their
Jupyter session and in a Docusaurus build.

Resolves #1473

### Changes proposed

- Updates to the `convert_ipynb_to_mdx.py` script;
  - search and find all inline `script` attributes, and reformat them;
  - fix deprecation warning about showing the index when a pandas
    DataFrame is saved to Markdown.
- Updates to notebooks that were previously using React style `style`
  values.
  - Coin_flipping.ipynb
  - GMM_with_2_dimensions_and_4_components.ipynb
  - Hidden_Markov_model.ipynb
  - Hierarchical_modeling.ipynb
  - Hierarchical_regression.ipynb
  - Item_Response_Theory.ipynb
  - Neals_funnel.ipynb
  - Robust_Linear_Regression.ipynb
  - Sparse_Logistic_Regression.ipynb
  - Zero_inflated_count_data.ipynb

### Test plan

Before when tutorials had to use React style attributes.

After the proposed changes.

# Types of changes

- [x] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

# Checklist

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **[CONTRIBUTING](https://github.com/facebookresearch/beanmachine/blob/main/CONTRIBUTING.md)** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] The title of my pull request is a short description of the requested changes.