Rewrite multiplications of negative reals
Summary:
While working on compiling the GEP model I noticed that the BMG type system loses track of the sign of negative reals when multiplied by a positive real.

This is because BMG's multiplication node requires that all inputs be the same type and that the type be real, positive real or probability. This means that the output type is always the same as the input type.

Unfortunately this means that `R- * R-` is of type `R`, not `R+`, and `R- * R+` is of type `R` and not `R-`, because in both cases the type requirements fixer concludes that `R * R --> R` is the only valid choice.

This is unfortunate when working with log probs. If we have probability p and we would like the log prob of p squared then `2 * log(p)` is sensible, but BMG then believes that the result is real, not negative real.

We can fix this problem by changing BMG to allow mixed negative real and positive real inputs to multiplications, but until then we can work around the problem by detecting affected multiplications and inserting negate nodes appropriately to ensure that all multiplications of mixed R- and R+ are done as R+ inputs to the multiplication, and then we optionally insert a negate at the end to ensure that the total number of negate nodes inserted is even.

Reviewed By: gafter

Differential Revision: D35767778

fbshipit-source-id: 908e0eac56ba56a6421d609a6311f202f72a5216