Implemented MALA proposer in bean machine
Summary:
Implemented MALAProposer by subclassing the HMCProposer and limiting it to 1 leapfrog step. The HMCPRoposer has a required attribute called trajectory_length which is not required in MALA because it takes only 1 leapfrog step. Three design alternatives were considered in the overridden method: .__init__() .propose():
1. Make trajectory_length an optional parameter for the methods in HMCProposer.
2. Re-use the base constructor with a hardcoded arbitrary value for trajectory_length and handle it internally in the MALAProposer class.
3. Reuse code and reimplement methods to not require trajectory_length in MAlAProposer.
Option 3 was chosen to not break future changes and / or explicit design choices in HMCProposer.

Differential Revision: D37008050

fbshipit-source-id: b99e7c77ebd642049e1cbdd74b032910c4d8751e