Mini BMG output from compiler (#1494)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1494

Neal's experiment with a mini version of BMG will have a JSON-based serialization format for graphs; in this diff I add a private `_to_mini` method to `BMGInference` which accumulates a graph and then dumps it out in this JSON format without otherwise rewriting the graph or checking for correctness.

This is intended as an expedient way to get end-to-end testing of this experiment sooner rather than later.

Reviewed By: gafter

Differential Revision: D37157629

fbshipit-source-id: 15534f428a96a11ee20d45a914e44247cc6ce4b7