minibmg: Containers: a C++ utility that can help keep client-specific fields out of the Graph API
Summary:
In order to avoid placing client-specific and derived data into a data structure, you can use this utility to make the data structure extensible by its clients.  Under the hood the data structure maintains a map of client data.

This is a generic data structure, so probably deserves to be migrated into a more generic library.  I would not be surprised to learn something like this already exists, but I don't know where to find it.  In any case, it is tiny.

This was motivated by reviewing D37098952.  It moves some data structures used by inference into the graph.  I have been trying to move such things out of the graph (e.g. in D37120689 (https://github.com/facebookresearch/beanmachine/commit/dc1fe6b3a19cc68f95afc35348699ed9678ea13a)) and at the moment these two directions are at odds with each other.  This utility may be a way to bring them together.

The design of this is loosely based on the idea of thread locals.

Reviewed By: rodrigodesalvobraz

Differential Revision: D37186542

fbshipit-source-id: cc3db5b76087d5c8b67d704e2ddbf5b92acea934