Improving the profiler report
Summary:
The internal deterministic profiler of BMG is very useful in accounting for the time spent by the profiled functions. It produces a JSON object that describes the performance run and the various measurements. There was an issue in the JSON processor in that it creates a tree whose root does not correspond to any profiling event. This creates an issue in that the total time at the end was reported as "unattributed time". This now is fixed.
The Diff consists of a simple change to account for the JSON tree root in file profiler.py under the compiler branch of the tree, and corresponding changes in two test files for the profiler code.

Reviewed By: ericlippert

Differential Revision: D36984640

fbshipit-source-id: b418f48a35bd6a2fecf7c2f1e87d401f5940c5eb