Introduce flag for Graph initialization
Summary: Introducing a flag for preventing multiple initialization of graph. This is needed because MH currently assumes Graph is not initialized and invokes `Graph::initialize()`. This will incur in multiple and unnecessary Graph initializations of MH is applied more than once to the same Graph instance. We could instead change MH's assumption to Graph being already initialized, but that would require the user remembering to initialize the Graph before running MH, which is error-prone.

Reviewed By: gafter

Differential Revision: D37099045

fbshipit-source-id: aa86ab50b2affe28b494b947528b13af5a6afe5b