Rename support functions and variables to be more precise.
Summary:
With the move of functions from MH to Graph, there was a name collision of functions both called `compute_support`, but with one of them returning node ids and the other returning node pointers. The latter received a temporary name to avoid the collision, but now the former has been renamed `compute_support_node_ids` while the latter is left as `compute_support`.

Once that change took place, I also renamed functions and data members to be more precise about being ordered and containing support nodes or node ids.

Differential Revision: https://www.internalfb.com/diff/D37162579?entry_point=27

fbshipit-source-id: e5a4839d1da4b3ee00397de8623d98ffe9e98064