Give more descriptive name to `Graph::initalize`
Summary: `Graph::initialize` creates multiple intermediate data structures for the graph so processing in evaluation and inference is faster. So make that function more clear, I've renamed it `Graph::ensure_evaluation_and_inference_readiness`. The "ensure" refers to the fact that the these data structures are built only the first time around. After that the call is simply ensuring this is the case.

Reviewed By: gafter

Differential Revision: D37164472

fbshipit-source-id: 2fd0839bf52438f98bffdc49657d7a6c900747fd