BMG New Profiler (#1460)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1460

This diff includes the class declaration and implementation of the new BMG profiler. It lives as a self-contained class manifested by a single object accessible from anywhere in the BMG code (or any other code for that matter). The code is isolated from the rest of the BMG code to promote maintainability and usability. While it is possible to generate more than one profiler object if the programmer so desires, and both objects will work without interference, there will be little utility in doing so and it is not recommended.

Reviewed By: ericlippert

Differential Revision: D36616797

fbshipit-source-id: 2262c3dcdfcbd0d64f379b7bf1b83be17e2130b9