Remove log1mexp "hint" workaround (#1480)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1480

If we have a log probability `x` in hand then `log(1-exp(x))` is the log probability of the inverse of `x`.  This `log1mexp` operation is therefore common and we would like to have a specific node in BMG for it -- and we do.

However, oddly enough torch does not have a dedicated `log1mexp` operation. Previous to this diff, to test generation of the `log1mexp` node by the compiler we recognized a "hint" method in our own utility library and generated a `log1mexp` node if we saw a stochastic call to that hint.

This is obviously not ideal; the better idea is to simply recognize models that contain `log(1-exp(x))` (or `(1-x.exp()).log()` or any other equivalent expression) and automatically convert it to a `log1mexp` node if `x` is a negative real.

I've done so, but unfortunately that has introduced a small problem.

Models often have arithmetic in them that you and I know always produce a log probability, but the type checker is insufficiently smart to know that. See comments in `to_neg_real` test for an example. In some of these cases we automatically insert an unsafe narrowing conversion from real to negative real.

This change causes us to no longer do so; I have not yet investigated why not, and I've disabled the failing test.

Reviewed By: feynmanliang

Differential Revision: D35800566

fbshipit-source-id: a6bbd3a2c849c5df923f1b8e30a64a0a54c34c2c