Rename support functions and variables to be more precise.
Summary:
With the move of functions from MH to Graph, there was a name collision of functions both called `compute_support`, but with one of them returning node ids and the other returning node pointers. The latter received a temporary name to avoid the collision, but now the former has been renamed `compute_support_node_ids` while the latter is left as `compute_support`.

Once that change took place, I also renamed functions and data members to be more precise about being ordered and containing support nodes or node ids.

Reviewed By: gafter

Differential Revision: D37162579

fbshipit-source-id: 6c3aebc1cf7e6accfcf2448acb03a331bdc58233