Implement Node datastructure (#1514)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1514

Background:
We are building Bayesian Additive Regression Trees (BART) as an experimental causal inference model in beanmachine. Details of the project can be found in https://docs.google.com/document/d/11nkB6UTGpvQBEC2yBjfgwAr8VabTlD7R9XufGQG0EvI/edit?usp=sharing and the proposed design can be found in the draft design document: https://docs.google.com/document/d/1o3J7yobDF0M9E27Y0tP2889fycmemXUZbHE5cebRqzs/edit?usp=sharing.

In this diff:
We have implemented the nodes of the tree.
Each node contains a set of rules used for conditioning or filtering the input data.  ```BaseNode``` class keeps track of left and right children and can condition / filter input based on its rules. ```LeafNode``` encapsulates the terminal nodes and can make a prediction based on its value. For a given input point, the prediction of the leaf node it is filtered to constitutes a prediction of the tree. It also contains the logic to "grow" which is the action of converting it to an internal node by growing terminal children based on some splitting rule.
```SplitNode``` encapsulates the internal nodes. It contains the logic of "pruning" which is the action of converting an internal node to a terminal node by removing its children, who must be leaves.

Reviewed By: feynmanliang

Differential Revision: D37478652

fbshipit-source-id: 444c931e68baba1b4a697da045295d3f082de78a