add copy node for parents and children of subgraph
Summary:
We want the subgraph to be fully independent without depending on any parent or children nodes from the graph.

This means that we need to copy the value of the parent and children nodes and replicate them within the subgraph.

For example, let's say our graph looks like

{F698197323}
and we want to marginalize "coin".

Both n1 and n2 are children of the MarginalizedDistribution, but the value of n2 depends on n1.

We add a COPY node here for the children, so it looks like
{F698197737}

To quote Rodrigo:
We want the marginalized graph to be a marginalized distribution D with two children n1, n2. D is a node that "remembers" the sub-graph all the way up to n1, n2 and should be defined such that D(n1,n2).log_prob = log \sum_coin Bernoulli.prob(coin) * Normal.prob(n1, coin) * Normal.prob(n2, n1).

We do need to keep some sort of variable between the two normals inside the sub-graph. Let's call that n1' (note the prime). We can then do the same for n2 (and have a n2'). So when node D is required to compute its log prob for a certain assignment of (n1, n2), all it needs to do is to plug those values in the sub-graph variables (n1', n2') and use normal BMG functions to compute the log prob.

There is a copy of those variables inside the sub-graph, and the marginalized distribution D knows about the n1 ↔ n1' and n2 ↔ n2' correspondence and uses that to perform its tasks.

Differential Revision: D33952498

fbshipit-source-id: 1833e3f0867cb34a67fc63e0d5152ae0bf210fd5