refactor marginalize graph for readability (#1522)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1522

This diff is to address comments left in Diffs 6-8.
Due to the big refactor in Diff 8 D34287403, it would have been a lot of work to update the diffs prior to that, since I would need to redo everything for the refactor.

I have gone back to Diffs 6-8 and resolved the comments which I've addressed. I've also copy and pasted all of the relevant comments so it'll be easier to review.

In summary, this diff
1. splits `marginalize_graph` into many helper functions
2. removes "COPY" node and instead uses
a. CONSTANT nodes for parents in the subgraph
b. SAMPLE nodes for stochastic children of MarginalDistribution
3. adds a test with upstream parents

Reviewed By: rodrigodesalvobraz

Differential Revision: D34718130

fbshipit-source-id: 8b4778995eb35a0a14d8b78a9fbac5e1cc74d4f7