NUTS generalized u-turn condition
Summary:
update NUTS as described in section A.4.2 of "A Conceptual Introduction to Hamiltonian Monte Carlo" by Michael Betancourt
https://arxiv.org/pdf/1701.02434.pdf

The main updates are to the `compute_no_turn` function.

### Original
The original no u-turn condition was:

{F705538398}
where
- `momentum` corresponds to "p"
- `position` corresponds to "q"
- `left` corresponds to "-"
- `right` corresponds to "+"

### New
The updated no u-turn condition is
{F705538538}
where
{F705538857}
and
{F705538907}

We currently are only supporting an identity mass matrix, so M^{-1} can be ignored.
Similar to the original implementation, `momentum` corresponds with "p".
However, $\rho(t)$ is approximated using a sum of all momentums along the trajectory. I introduce a new variable named `momentum_sum`, which stores the sum of all the momentums starting from time 0.

Reviewed By: mootaz77, ToddSmall

Differential Revision: D30712345

fbshipit-source-id: 615eb7554e8dbb26c001e2d8eb3477ee08253c30