mass matrix adaptation with windows
Summary:
NUTS and HMC have a mass matrix that is equivalent to the covariance of the samples during the adaptation. This diff creates `DiagonalCovarianceComputer` to compute the covariance using Welford's online algorithm, and the diff also contains `WindowedMassMatrixAdapter` to call the covariance algorithm appropriately.

This classes are not attached to the NUTS algorithm yet.

`DiagonalCovarianceComputer` uses Welford's online algorithm to efficiently compute the diagonal covariance matrix.
References:
https://docs.pyro.ai/en/1.7.0/_modules/pyro/ops/welford.html
https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Welford's_online_algorithm

`WindowedMassMatrixAdapter` calls `DiagonalCovarianceComputer` to compute the covariance of samples in windows as described in Stan's documentation:
https://mc-stan.org/docs/2_27/reference-manual/hmc-algorithm-parameters.html
In summary, the mass matrix is updated in windows, with the size of the windows determined by the total number of samples.
Stan's calculation of window sizes: https://github.com/stan-dev/stan/blob/develop/src/stan/mcmc/windowed_adaptation.hpp

Reviewed By: mootaz77, ToddSmall

Differential Revision: D31662165

fbshipit-source-id: e81460a96b0cb5734e3245b6721891528b6d29fe