NUTS multinomial sampling
Summary:
Add option for multinomial sampling rather than uniform as explained in A.2 of https://arxiv.org/pdf/1701.02434.pdf

In the original NUTS algorithm, when combining the left and right tree in `build_tree` and selecting the new sample value, we would uniformly choose between all the values along the left and right trajectory that were "valid" (value < slice variable).

With multinomial sampling, we no longer use the slice variable. We instead choose the value based on a multinomial distribution

{F705620224}

where z is the sample value and \pi is the Hamiltonian.
In practice, we use the log of the hamiltonian

Differential Revision: D32931760

fbshipit-source-id: e409ef4de6cd6cd876f64a2a66567571d4790666