robust uturn
Summary:
As described in [a Stan forum](https://discourse.mc-stan.org/t/nuts-misses-u-turns-runs-in-circles-until-max-treedepth/9727), the NUTS stopping condition doesn't work properly in some instances.

This diff adds additional u-turn checks between subtrees to prevent this behavior. It adds a u-turn check between the leftmost node of the right tree and the left tree, as well as the rightmost node of the left tree and the right tree.

Reviewed By: mootaz77

Differential Revision: D32954189

fbshipit-source-id: a887f13dc33714b32d060ccc30c8816d2adf3c0e